requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# IBKR Integration
ib-insync==0.9.86

# AI/ML Libraries
torch==2.1.1
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
tensorflow==2.15.0

# API & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
httpx==0.25.2
aiohttp==3.9.1

# Message Queue & Cache
redis==5.0.1
kafka-python==2.0.2
celery==5.3.4

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Network Analysis
networkx==3.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==22.0.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
